move logger to pkg
