+++ cmd/server/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"go-pass-keeper/internal/app"
 	"go-pass-keeper/internal/grpcserver/config"
-	"go-pass-keeper/internal/logger"
+	"go-pass-keeper/pkg/logger"
 )
 
 // функция main вызывается автоматически при запуске приложения
