+++ internal/app/app.go
@@ -7,10 +7,10 @@ import (
 	"go-pass-keeper/internal/grpcserver"
 	"go-pass-keeper/internal/grpcserver/config"
 	interceptors "go-pass-keeper/internal/grpcserver/interceptors"
-	"go-pass-keeper/internal/logger"
 	"go-pass-keeper/internal/services"
 	"go-pass-keeper/internal/storage"
 	"go-pass-keeper/internal/token"
+	"go-pass-keeper/pkg/logger"
 	"os"
 	"os/signal"
 	"syscall"
