+++ internal/services/user_test.go
@@ -4,10 +4,10 @@ import (
 	"context"
 	"errors"
 	"go-pass-keeper/internal/grpcserver/config"
-	"go-pass-keeper/internal/logger"
 	"go-pass-keeper/internal/models"
 	"go-pass-keeper/internal/storage"
 	"go-pass-keeper/internal/token"
+	"go-pass-keeper/pkg/logger"
 	pb "go-pass-keeper/pkg/proto"
 	"testing"
 	"time"
